Fix making symlinks when dir not exists

+++ symlink_dotfiles.sh
@@ -5,10 +5,16 @@ do
 		cp $PWD/$i.sample $PWD/$i
 		echo "copied $PWD/$i.sample $PWD/$i"
 	fi
-	if [ -f ~/$i ]; then
-		mv ~/$i ~/$i.original
-		echo "moved ~/$i ~/$i.original"
+	if [ ! -d `dirname ~/$i` ]; then
+		mkdir -p `dirname ~/$i`
+		echo "created `dirname ~/$i`"
 	fi
-	ln -s $PWD/$i ~/$i
-	echo "linked $PWD/$i ~/$i"
+	if [ ! -L ~/$i ]; then
+		if [ -f ~/$i ]; then
+			mv ~/$i ~/$i.original
+			echo "moved ~/$i ~/$i.original"
+		fi
+		ln -s $PWD/$i ~/$i
+		echo "linked $PWD/$i ~/$i"
+	fi
 done
